PromptEngineering/LLM_tests_framework/src/frameworks/deepeval_metrics.py
from deepeval import evaluate, TestCase, LLMTestCase
from deepeval.metrics import (
    HallucinationMetric,
    AnswerRelevancyMetric,
    ContextualPrecisionMetric,
    ContextualRecallMetric,
    FaithfulnessMetric,
    BiasMetric,
    ToxicityMetric,
    RAGASMetric
)
from deepeval.dataset import TestDataset
from datetime import datetime
import logging
import json
from typing import Dict, List, Any
from concurrent.futures import ThreadPoolExecutor

class DeepEvalMetrics:
    def __init__(self, config_path: str):
        self.config = self._load_config(config_path)
        self.setup_logging()
        self.initialize_metrics()
        self.history = []

    def setup_logging(self):
        """Configure logging for test results and errors"""
        logging.basicConfig(
            filename=f'llm_testing_{datetime.now().strftime("%Y%m%d")}.log',
            level=logging.INFO,
            format='%(asctime)s - %(levelname)s - %(message)s'
        )

    def initialize_metrics(self):
        """Initialize all available metrics"""
        self.metrics = {
            'hallucination': HallucinationMetric(),
            'relevancy': AnswerRelevancyMetric(),
            'contextual_precision': ContextualPrecisionMetric(),
            'contextual_recall': ContextualRecallMetric(),
            'faithfulness': FaithfulnessMetric(),
            'bias': BiasMetric(),
            'toxicity': ToxicityMetric(),
            'ragas': RAGASMetric()
        }

    def _load_config(self, config_path: str) -> Dict:
        """Load configuration from JSON file"""
        with open(config_path, 'r') as file:
            return json.load(file)

    async def continuous_evaluation(self, interval_minutes: int = 60):
        """Run continuous evaluation at specified intervals"""
        while True:
            try:
                results = await self.run_full_evaluation()
                self.history.append({
                    'timestamp': datetime.now(),
                    'results': results
                })
                self.analyze_trends()
                await asyncio.sleep(interval_minutes * 60)
            except Exception as e:
                logging.error(f"Continuous evaluation error: {str(e)}")

    def run_full_evaluation(self) -> Dict[str, Any]:
        """Run comprehensive evaluation across all models and metrics"""
        results = {}
        with ThreadPoolExecutor(max_workers=self.config['max_workers']) as executor:
            for model_name, model_config in self.config['models'].items():
                future = executor.submit(self.evaluate_model_all_metrics, model_name, model_config)
                results[model_name] = future.result()
        return results

    def evaluate_model_all_metrics(self, model_name: str, model_config: Dict) -> Dict[str, float]:
        """Evaluate a single model across all available metrics"""
        test_cases = self.prepare_test_cases(model_config['test_cases'])
        results = {}

        for metric_name, metric in self.metrics.items():
            score = evaluate(
                model=model_config['model'],
                test_cases=test_cases,
                metrics=[metric]
            )
            results[metric_name] = score

        return results

    def prepare_test_cases(self, test_config: Dict) -> List[TestCase]:
        """Prepare test cases from configuration"""
        return [TestCase(**case) for case in test_config]

    def analyze_trends(self):
        """Analyze historical performance trends"""
        if len(self.history) < 2:
            return

        current = self.history[-1]['results']
        previous = self.history[-2]['results']

        significant_changes = self._detect_significant_changes(current, previous)
        if significant_changes:
            self._send_alert(significant_changes)

    def _detect_significant_changes(self, current: Dict, previous: Dict) -> List[str]:
        """Detect significant changes in model performance"""
        threshold = self.config['alert_threshold']
        changes = []

        for model in current:
            for metric in current[model]:
                current_value = current[model][metric]
                previous_value = previous[model][metric]

                if abs(current_value - previous_value) > threshold:
                    changes.append(f"{model} {metric}: {previous_value:.2f} -> {current_value:.2f}")

        return changes

    def _send_alert(self, changes: List[str]):
        """Send alerts for significant performance changes"""
        logging.warning(f"Significant changes detected: {', '.join(changes)}")

    def generate_report(self) -> Dict[str, Any]:
        """Generate comprehensive evaluation report"""
        latest_results = self.history[-1]['results'] if self.history else None
        if not latest_results:
            return {}

        report = {
            'timestamp': datetime.now().isoformat(),
            'summary': self._generate_summary(latest_results),
            'detailed_results': latest_results,
            'trends': self._analyze_historical_trends(),
            'recommendations': self._generate_recommendations(latest_results)
        }

        return report

    def _generate_summary(self, results: Dict) -> Dict[str, float]:
        """Generate summary statistics from results"""
        summary = {}
        for model in results:
            summary[model] = {
                'average_score': sum(results[model].values()) / len(results[model]),
                'best_metric': max(results[model].items(), key=lambda x: x[1])[0],
                'worst_metric': min(results[model].items(), key=lambda x: x[1])[0]
            }
        return summary

    def _analyze_historical_trends(self) -> Dict[str, Any]:
        """Analyze historical performance trends"""
        if len(self.history) < 2:
            return {}

        trends = {}
        for model in self.history[-1]['results']:
            trends[model] = {
                'improvement': self._calculate_improvement(model),
                'stability': self._calculate_stability(model)
            }
        return trends

    def _calculate_improvement(self, model: str) -> float:
        """Calculate improvement over time for a model"""
        # Implementation here
        pass

    def _calculate_stability(self, model: str) -> float:
        """Calculate stability over time for a model"""
        # Implementation here
        pass

    def _generate_recommendations(self, results: Dict) -> List[str]:
        """Generate recommendations based on results"""
        # Implementation here
        pass